{aq_variance,firstpass,ratectrl}.c: quiet -Wshorten warnings
ported from libvpx:

Added casts to remove warnings:

In regards to the safety of these casts they are of two types:-

- Normalized bits per (16x16) MB stored in a 32 bit int (This is safe as
  bits per MB even with << 9 normalization cant overflow 32 bits. Even
  raw 12 bits hdr source even would only be  29 bits :- (4+4+12+9) and
  the encoder imposes much stricter limits than this on max bit rate.

- Cast as part of variance calculations.  There is an internal cast up
  to 64 bit for the Sum X Sum calculation, but after normalization
  dividing by the number of points the result will always be <= the SSE
  value.

BUG=aomedia:445

+++ av1/encoder/aq_variance.c
@@ -169,8 +169,8 @@ static unsigned int block_variance(const AV1_COMP *const cpi, MACROBLOCK *x,
     aq_variance(x->plane[0].src.buf, x->plane[0].src.stride, av1_all_zeros, 0,
                 bw, bh, &sse, &avg);
 #endif  // CONFIG_HIGHBITDEPTH
-    var = sse - (((int64_t)avg * avg) / (bw * bh));
-    return ((uint64_t)var * 256) / (bw * bh);
+    var = sse - (unsigned int)(((int64_t)avg * avg) / (bw * bh));
+    return (unsigned int)((uint64_t)var * 256) / (bw * bh);
   } else {
 #if CONFIG_HIGHBITDEPTH
     if (xd->cur_buf->flags & YV12_FLAG_HIGHBITDEPTH) {
@@ -185,7 +185,7 @@ static unsigned int block_variance(const AV1_COMP *const cpi, MACROBLOCK *x,
     var = cpi->fn_ptr[bs].vf(x->plane[0].src.buf, x->plane[0].src.stride,
                              av1_all_zeros, 0, &sse);
 #endif  // CONFIG_HIGHBITDEPTH
-    return ((uint64_t)var * 256) >> num_pels_log2_lookup[bs];
+    return (unsigned int)((uint64_t)var * 256) >> num_pels_log2_lookup[bs];
   }
 }
